feat(client): add client struct
