+++ src/net/src/lib.rs
@@ -17,17 +17,18 @@
  * limitations under the License.
  */
 
-mod base_cmd;
 pub mod cmd_group_client;
 pub mod cmd_kv;
+pub mod handle;
+pub mod tcp;
+
+mod base_cmd;
+mod client;
 mod cmd_table;
 mod error;
-pub mod handle;
 mod resp;
-pub mod tcp;
 mod unix;
 
-use crate::resp::RespProtocol;
 use crate::tcp::TcpServer;
 use async_trait::async_trait;
 use std::error::Error;
@@ -35,84 +36,6 @@ use std::error::Error;
 #[async_trait]
 pub trait ServerTrait: Send + Sync + 'static {
     async fn run(&self) -> Result<(), Box<dyn Error>>;
-}
-
-#[async_trait]
-pub trait StreamTrait: Send + Sync {
-    async fn read(&mut self, buf: &mut [u8]) -> Result<usize, std::io::Error>;
-    async fn write(&mut self, data: &[u8]) -> Result<usize, std::io::Error>;
-}
-
-pub struct Client {
-    stream: Box<dyn StreamTrait>,
-    // TODO: use &[Vec<u8>], need lifetime.
-    argv: Vec<Vec<u8>>,
-    // Client name.
-    name: Vec<u8>,
-    cmd_name: Vec<u8>,
-    key: Vec<u8>,
-    reply: RespProtocol,
-}
-
-impl Client {
-    pub fn new(stream: Box<dyn StreamTrait>) -> Self {
-        Self {
-            stream,
-            argv: Vec::default(),
-            name: Vec::default(),
-            cmd_name: Vec::default(),
-            key: Vec::default(),
-            reply: RespProtocol::new(),
-        }
-    }
-
-    pub async fn read(&mut self, buf: &mut [u8]) -> Result<usize, std::io::Error> {
-        self.stream.read(buf).await
-    }
-
-    pub async fn write(&mut self, data: &[u8]) -> Result<usize, std::io::Error> {
-        self.stream.write(data).await
-    }
-
-    pub fn set_argv(&mut self, argv: &[Vec<u8>]) {
-        self.argv = argv.to_vec()
-    }
-
-    pub fn argv(&self) -> &[Vec<u8>] {
-        &self.argv
-    }
-
-    pub fn set_name(&mut self, name: &[u8]) {
-        self.name = name.to_vec()
-    }
-
-    pub fn name(&self) -> &[u8] {
-        &self.name
-    }
-
-    pub fn set_cmd_name(&mut self, name: &[u8]) {
-        self.cmd_name = name.to_vec()
-    }
-
-    pub fn cmd_name(&self) -> &[u8] {
-        &self.cmd_name
-    }
-
-    pub fn set_key(&mut self, key: &[u8]) {
-        self.key = key.to_vec()
-    }
-
-    pub fn key(&self) -> &[u8] {
-        &self.key
-    }
-
-    pub fn reply_mut(&mut self) -> &mut RespProtocol {
-        &mut self.reply
-    }
-
-    pub fn take_reply(&mut self) -> RespProtocol {
-        std::mem::replace(&mut self.reply, RespProtocol::new())
-    }
 }
 
 pub struct ServerFactory;
